requirements.txt
django==2.1
django-mptt==0.9.1
